+++ 1k/fetch.cmake
@@ -41,7 +41,7 @@ endfunction()
 # param package_name
 function(_1kfetch_dist package_name)
     set(_prebuilt_root ${CMAKE_CURRENT_LIST_DIR}/_x)
-    if(_1KFETCH_DIST_UPGRADE AND IS_DIRECTORY ${_prebuilt_root})
+    if(_1KFETCH_UPGRADE AND IS_DIRECTORY ${_prebuilt_root})
         file(REMOVE_RECURSE ${_prebuilt_root})
     endif()
     if(NOT IS_DIRECTORY ${_prebuilt_root})