+++ core/2d/TextFieldTTF.h
@@ -39,7 +39,8 @@ NS_AX_BEGIN
 class TextFieldTTF;
 
 /**
- * A input protocol for TextField.
+ * A input protocol for TextField. 
+ * !!!DEPRECATED since axmol-2.1.3
  */
 class AX_DLL TextFieldDelegate
 {